requirements.txt
appdirs==1.4.0
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
Faker==8.11.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.3
image==1.5.33
importlib-metadata==4.8.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==16.8
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.1.10
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
zipp==3.5.0
